Refine file structure
